heat2030: move GPlanning to common
