+++ src/climatevision/generator/heat2030/energy_general/g_storage.py
@@ -55,45 +55,3 @@ class GStorage(G):
             power_to_be_installed=power_to_be_installed,
             ratio_wage_to_emplo=ratio_wage_to_emplo,
         )
-
-
-@dataclass(kw_only=True)
-class GPlanning(G):
-    pct_of_wage: float
-    ratio_wage_to_emplo: float
-
-    @classmethod
-    def calc(cls, inputs: Inputs) -> "GPlanning":
-        entries = inputs.entries
-        ass = inputs.ass
-        fact = inputs.fact
-
-        invest = (
-            fact("Fact_H_P_planning_cost_basis")
-            + fact("Fact_H_P_planning_cost_per_capita") * entries.m_population_com_2018
-        )
-        invest_pa = invest / entries.m_duration_target
-
-        invest_com = invest
-        invest_pa_com = invest_pa
-
-        pct_of_wage = ass("Ass_H_G_planning_cost_pct_of_wage")
-        cost_wage = invest / fact("Fact_H_P_planning_duration")
-
-        ratio_wage_to_emplo = ass("Ass_T_C_yearly_costs_per_planer")
-        demand_emplo = div(cost_wage, ratio_wage_to_emplo)
-        demand_emplo_new = demand_emplo
-        demand_emplo_com = demand_emplo_new
-
-        return cls(
-            cost_wage=cost_wage,
-            demand_emplo=demand_emplo,
-            demand_emplo_com=demand_emplo_com,
-            demand_emplo_new=demand_emplo_new,
-            invest=invest,
-            invest_com=invest_com,
-            invest_pa=invest_pa,
-            invest_pa_com=invest_pa_com,
-            pct_of_wage=pct_of_wage,
-            ratio_wage_to_emplo=ratio_wage_to_emplo,
-        )
